fixed illegal move bug and bad move bug
Fixed illegal move bug (it was returning an illegal move after 0 nodes searched, I've made it return a known legal move if it searches 0 moves).

Fixed a bug where it was making bad moves once every few games. This was caused by it returning a draw score for nodes that weren't searched due to running out of time. I changed this to -MATE_SCORE and it fixed the issue.