Temporary fix for disconnection bug.
Once every few thousand games, Raven will disconnect in cutechess. When this happens, Raven searches very deeply in a very small amount of time and produces the same PV and score for each depth. If the search has reached depth 25 within 10 milliseconds, Raven now returns the result to avoid the crash. This is a temporary fix and I'll look into the cause of the crashing.